fix the types
